refactor(settings/models): 提取 AddModelDialog 为独立组件
