+++ src/renderer/pages/settings/models/components/AddModelDialog.tsx
@@ -1,0 +1,268 @@
+import { Button, Checkbox, Form, Input, InputNumber, Message, Select, Spin, Tag } from '@arco-design/web-react';
+import { Download, PreviewOpen } from '@icon-park/react';
+import React, { useEffect, useMemo, useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { ipcBridge } from '@/common';
+import type { ScodeConfig } from '@/common/ipcBridge';
+import { buildCustomModelAlias, type ScodeCustomModelProvider } from '@/common/scodeConfig';
+import AionModal from '@/renderer/components/base/AionModal';
+import { editableModelFromEntry, findModelEntry, presetValueForProvider, sanitizeProviderId, PROVIDER_PRESETS } from '../utils';
+import type { EditableModel, EditingModelTarget } from '../types';
+
+export default function AddModelDialog({ visible, onClose, onSubmit, existingProviderIds, existingModelIds, config, editingTarget }: IAddModelDialogProps) {
+  const { t } = useTranslation();
+  const [form] = Form.useForm();
+  const [saving, setSaving] = useState(false);
+  const [showApiKey, setShowApiKey] = useState(false);
+  const [fetchingModels, setFetchingModels] = useState(false);
+  const [providerModels, setProviderModels] = useState<string[]>([]);
+  const providerPreset = Form.useWatch('providerPreset', form);
+  const selectedPreset = useMemo(() => PROVIDER_PRESETS.find((item) => item.value === providerPreset), [providerPreset]);
+  const isEditing = Boolean(editingTarget);
+  const modelOptions = useMemo(() => providerModels.map((model) => ({ label: model, value: model })), [providerModels]);
+
+  useEffect(() => {
+    if (!visible) return;
+    if (editingTarget) {
+      const entry = findModelEntry(config, editingTarget.modelId);
+      const editableModel = editableModelFromEntry(config, editingTarget.modelId);
+      const input = entry?.input || [];
+      form.setFieldsValue({
+        providerPreset: presetValueForProvider(editingTarget.provider),
+        providerId: editingTarget.provider.id,
+        baseUrl: editingTarget.provider.baseUrl,
+        apiKey: editingTarget.provider.apiKey,
+        modelId: editableModel.id,
+        supportsTools: entry?.supports_tools !== false,
+        supportsVision: input.includes('image'),
+        supportsReasoning: Boolean(entry?.supports_reasoning),
+        inputContext: entry?.context?.input,
+        outputContext: entry?.context?.output,
+      });
+      setShowApiKey(false);
+      setProviderModels([]);
+      return;
+    }
+
+    form.setFieldsValue({
+      providerPreset: 'custom',
+      providerId: 'custom-openai',
+      baseUrl: '',
+      apiKey: '',
+      modelId: '',
+      supportsTools: true,
+      supportsVision: false,
+      supportsReasoning: false,
+    });
+    setShowApiKey(false);
+    setProviderModels([]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [form, visible]);
+
+  useEffect(() => {
+    if (isEditing) return;
+    const preset = PROVIDER_PRESETS.find((item) => item.value === providerPreset);
+    if (!preset) return;
+    form.setFieldsValue({
+      providerId: preset.providerId,
+      baseUrl: preset.baseUrl,
+    });
+    setProviderModels([]);
+  }, [form, isEditing, providerPreset]);
+
+  const handleFetchProviderModels = async () => {
+    const baseUrl = String(form.getFieldValue('baseUrl') || '').trim();
+    const apiKey = String(form.getFieldValue('apiKey') || '').trim();
+    if (!baseUrl) {
+      Message.warning(t('settings.pleaseEnterBaseUrl'));
+      return;
+    }
+    if (selectedPreset?.apiKeyRequired !== false && !apiKey) {
+      Message.warning(t('settings.pleaseEnterApiKey'));
+      return;
+    }
+
+    setFetchingModels(true);
+    try {
+      const res = await ipcBridge.mode.fetchModelList.invoke({
+        base_url: baseUrl,
+        api_key: apiKey,
+        platform: 'custom',
+      });
+      const models =
+        res.data?.mode
+          ?.map((item) => (typeof item === 'string' ? item : item.id))
+          .map((item) => item.trim())
+          .filter(Boolean) || [];
+
+      if (!res.success || models.length === 0) {
+        setProviderModels([]);
+        Message.warning(res.msg || t('settings.fetchModelListUnsupported'));
+        return;
+      }
+
+      setProviderModels(Array.from(new Set(models)));
+      Message.success(t('settings.fetchModelListSuccess', { count: models.length }));
+    } catch {
+      setProviderModels([]);
+      Message.warning(t('settings.fetchModelListUnsupported'));
+    } finally {
+      setFetchingModels(false);
+    }
+  };
+
+  const handleSubmit = async () => {
+    const values = await form.validate();
+    const providerId = sanitizeProviderId(values.providerId);
+    const modelId = String(values.modelId || '').trim();
+    const modelAlias = buildCustomModelAlias(providerId, modelId);
+    if (!providerId) {
+      Message.error(t('settings.sudocodeModel.providerIdInvalid', 'Provider ID 无效'));
+      return;
+    }
+    if (existingProviderIds.some((item) => item === providerId && item !== editingTarget?.provider.id)) {
+      Message.error(t('settings.sudocodeModel.providerIdExists', 'Provider ID 已存在，请换一个名称'));
+      return;
+    }
+    if (existingModelIds.some((item) => item === modelAlias && item !== editingTarget?.modelId)) {
+      Message.error(t('settings.sudocodeModel.modelIdExists', '模型名称已存在，请换一个名称'));
+      return;
+    }
+
+    const input = ['text'];
+    if (values.supportsVision) {
+      input.push('image');
+    }
+    const nextModel: EditableModel = {
+      id: modelId,
+      name: modelId,
+      input,
+      supportsTools: values.supportsTools,
+      supportsReasoning: values.supportsReasoning,
+      inputContext: values.inputContext,
+      outputContext: values.outputContext,
+    };
+
+    const models = editingTarget ? editingTarget.provider.modelIds.map((item) => (item === editingTarget.modelId ? nextModel : editableModelFromEntry(config, item))) : [nextModel];
+
+    setSaving(true);
+    try {
+      await onSubmit(
+        {
+          providerId,
+          baseUrl: values.baseUrl,
+          apiKey: values.apiKey || '',
+          models,
+        },
+        editingTarget?.provider.id,
+        editingTarget?.modelId,
+        modelAlias
+      );
+      onClose();
+    } finally {
+      setSaving(false);
+    }
+  };
+
+  return (
+    <AionModal
+      visible={visible}
+      header={{ title: isEditing ? t('settings.sudocodeModel.editModelTitle', '编辑模型') : t('settings.addModel', '添加模型'), showClose: true }}
+      style={{ width: 760 }}
+      contentStyle={{ background: 'var(--bg-1)', padding: '20px 24px 16px', overflow: 'auto' }}
+      onCancel={onClose}
+      footer={null}
+    >
+      <div className='mb-4 flex items-center gap-2'>
+        <Tag bordered>{t('settings.sudocodeModel.openAIOnlyHint', '仅支持 OpenAI 兼容协议 API')}</Tag>
+      </div>
+      <Form form={form} layout='vertical'>
+        <Form.Item label={t('settings.sudocodeModel.providerLabel', '提供商')} field='providerPreset' rules={[{ required: true }]}>
+          <Select options={PROVIDER_PRESETS} />
+        </Form.Item>
+        <Form.Item label='Provider ID' field='providerId' rules={[{ required: true }]} extra={t('settings.sudocodeModel.providerIdExtra', '写入 sudocode.json auth_modes.api-key 下的唯一 key。')}>
+          <Input
+            placeholder='custom-openai'
+            onBlur={(event) => {
+              form.setFieldValue('providerId', sanitizeProviderId(event.target.value));
+            }}
+          />
+        </Form.Item>
+        <Form.Item label={t('settings.sudocodeModel.baseUrlLabel', '接口地址')} field='baseUrl' rules={[{ required: true }]}>
+          <Input placeholder='https://api.example.com/v1' />
+        </Form.Item>
+        <Form.Item label='API Key' field='apiKey' rules={selectedPreset?.apiKeyRequired === false ? [] : [{ required: true }]}>
+          <Input
+            type={showApiKey ? 'text' : 'password'}
+            placeholder={selectedPreset?.apiKeyRequired === false ? t('settings.sudocodeModel.apiKeyOptionalPlaceholder', '本地服务可留空') : t('settings.sudocodeModel.apiKeyRequiredPlaceholder', '输入你的 API Key')}
+            suffix={<Button type='text' size='mini' icon={<PreviewOpen />} onClick={() => setShowApiKey((prev) => !prev)} />}
+          />
+        </Form.Item>
+        <Form.Item
+          label={
+            <div className='flex items-center justify-between gap-3 w-full'>
+              <span>{t('settings.sudocodeModel.modelIdLabel', '模型名称')}</span>
+              <Button size='mini' icon={<Download />} loading={fetchingModels} onClick={() => void handleFetchProviderModels()}>
+                {t('settings.fetchModelList')}
+              </Button>
+            </div>
+          }
+          field='modelId'
+          rules={[{ required: true }]}
+          extra={providerModels.length > 0 ? t('settings.modelListSelectOrManualHint') : t('settings.modelListFetchHint')}
+        >
+          <Select
+            showSearch
+            allowCreate
+            options={modelOptions}
+            loading={fetchingModels}
+            placeholder={t('settings.sudocodeModel.modelIdPlaceholder', '例如 gpt-4o 或 openai/gpt-4o')}
+            notFoundContent={fetchingModels ? <Spin size={16} /> : t('settings.modelListManualInputHint')}
+            filterOption={(inputValue, option) => {
+              const value = String((option as React.ReactElement<{ value?: string }>)?.props?.value || '');
+              return value.toLowerCase().includes(inputValue.toLowerCase());
+            }}
+          />
+        </Form.Item>
+        <div className='mb-2 text-14px font-600 text-foreground'>{t('settings.sudocodeModel.advancedConfig', '高级配置')}</div>
+        <div className='grid grid-cols-1 md:grid-cols-3 gap-3 mb-4'>
+          <Form.Item field='supportsTools' triggerPropName='checked'>
+            <Checkbox>{t('settings.sudocodeModel.supportsTools', '工具调用')}</Checkbox>
+          </Form.Item>
+          <Form.Item field='supportsVision' triggerPropName='checked'>
+            <Checkbox>{t('settings.sudocodeModel.supportsVision', '图片输入')}</Checkbox>
+          </Form.Item>
+          <Form.Item field='supportsReasoning' triggerPropName='checked'>
+            <Checkbox>{t('settings.sudocodeModel.supportsReasoning', '推理模式')}</Checkbox>
+          </Form.Item>
+        </div>
+        <div className='grid grid-cols-1 md:grid-cols-2 gap-4'>
+          <Form.Item label={t('settings.sudocodeModel.inputContextLabel', '输入上下文（K）')} field='inputContext'>
+            <InputNumber min={1} placeholder={t('settings.sudocodeModel.contextDefaultPlaceholder', '使用提供商默认值')} className='w-full' />
+          </Form.Item>
+          <Form.Item label={t('settings.sudocodeModel.outputContextLabel', '输出上下文（K）')} field='outputContext'>
+            <InputNumber min={1} placeholder={t('settings.sudocodeModel.contextDefaultPlaceholder', '使用提供商默认值')} className='w-full' />
+          </Form.Item>
+        </div>
+      </Form>
+      <div className='flex justify-end gap-2.5 mt-2.5'>
+        <Button onClick={onClose} className='px-5 min-w-20' style={{ borderRadius: 'var(--radius-md)' }}>
+          {t('common.cancel', '取消')}
+        </Button>
+        <Button type='primary' onClick={handleSubmit} loading={saving} className='px-5 min-w-20' style={{ borderRadius: 'var(--radius-md)' }}>
+          {t('common.save', '保存')}
+        </Button>
+      </div>
+    </AionModal>
+  );
+}
+
+interface IAddModelDialogProps {
+  visible: boolean;
+  onClose: () => void;
+  onSubmit: (provider: ScodeCustomModelProvider, previousProviderId?: string, previousModelId?: string, nextModelId?: string) => Promise<void>;
+  existingProviderIds: string[];
+  existingModelIds: string[];
+  config: ScodeConfig | null;
+  editingTarget: EditingModelTarget | null;
+}
